+++ .env.local
@@ -4,6 +4,7 @@ APP_KEY=base64:GrS4DCwWKC11OqYJxdoDQMkn3UTx8xsdavQVWOF87pI=
 # The port 80 is used because the puppeteer runs in a docker container
 APP_URL=http://127.0.0.1:80
 APP_ENV=local
+DEBUG=true
 
 PHOTOS_PATH=/var/www/html/storage/app/testing/photos
