+++ resources/views/archive/layout.blade.php
@@ -1,11 +1,11 @@
-@extends("archiviodocumenti.layout")
+@extends("layouts.app")
 
 @section("title", "Archivio Documenti")
 
 @section("navbar-link")
     <li class="nav-item">
         <a class="nav-link" href="{{ route("docs.index") }}">
-            Archivio Documenti
+            Archivio
         </a>
     </li>
      <li class="nav-item">
