+++ resources/views/archive/show.blade.php
